Simulating dispersal

The function simulates dispersal, for every square calculating the no. of individuals that disperse out of this square and the no. of individuals that disperse into this square.

Usage

disp(
  N_t,
  id,
  data_table,
  kernel,
  dens_dep,
  dlist,
  id_within,
  within_mask,
  border,
  max_dist,
  resolution,
  ncells_in_circle,
  cl = NULL
)

Arguments

N_t

integer matrix of population numbers at a single time step; NA stands for cells that are outside the study area

id

RasterLayer object (of the same size as N_t) with cell identifiers

data_table

matrix that contains information about all cells in current time points

kernel

a function defining dispersal kernel

dens_dep

character vector of length 1 specifying if the probability of settling in a target square is (case-sensitive):

"none" - fully random,

"K" - proportional to the carrying capacity of a target square,

"K2N" - density-dependent, i.e. proportional to the ratio of carrying capacity of a target square to the number of individuals already present in a target square

dlist

a list with identifiers of target cells at a specified distance from a focal cell

id_within

integer vector with identifiers of cells inside the study area

within_mask

logical matrix that specifies boundaries of the study area

border

character vector of length 1 defining how to deal with borders (case-sensitive):

"absorbing" - individuals that disperse outside the study area are removed from the population

"reprising" - cells outside the study area are not allowed as targets for dispersal

max_dist

a distance (in the same units as used in the raster id) specifying the maximum range at which identifiers of target dispersal cells are determined in advance (see initialise)

resolution

integer vector of length 1; spatial resolution of id raster

ncells_in_circle

numeric vector; number of cells on each distance

cl

if simulation is done in parallel, the name of a cluster object created by makeCluster

Value

The function returns a list that contains two matrices:

em - emigration matrix with the number of individuals that dispersed from each cell im - immigration matrix with the number of individuals that dispersed to each cell

Details

This is the function used by sim internally and is not intended to be called by the user. The parameters for this function are passed from a sim_data object created by initialise.

Dispersal distance is expressed in original spatial units of the SpatRaster provided to sim function (n1_map and K_map), however, it is internally converted to units of the simulation (i.e. to the size of a single cell) by calculating round(distance/resolution). Thus, if the selected dispersal distance is smaller than resolution/2, it effectively means that an individual does not disperse, i.e. it lands in the same cell as it was before. The dispersal rate (proportion of dispersing individuals) can be estimated from the dispersal kernel probability function by calculating the probability that the dispersal distance is greater than resolution/2.

Examples

# data preparation
library(terra)
#> terra 1.7.39

n1_small <- rast(system.file("input_maps/n1_small.tif", package = "rangr"))
K_small <- rast(system.file("input_maps/K_small.tif", package = "rangr"))

sim_data <- initialise(
  n1_map = n1_small,
  K_map = K_small,
  r = log(2),
  rate = 1 / 1e3
)
#> NaN values were found in input maps and replaced with NA (cells outside the study area)
#> Warning: sf required for evolution_status==2L

# disp
disp_output <- disp(
  N_t = sim_data$n1_map,
  id = raster::raster(sim_data$id),
  data_table = sim_data$data_table,
  kernel = sim_data$kernel,
  dens_dep = sim_data$dens_dep,
  dlist = sim_data$dlist,
  id_within = sim_data$id_within,
  within_mask = sim_data$within_mask,
  border = sim_data$border,
  max_dist = sim_data$max_dist,
  resolution = sim_data$resolution,
  ncells_in_circle = sim_data$ncells_in_circle
)

# immigration and emigration matrices
names(disp_output)
#> [1] "em" "im"